pyproject.toml
[project]
name = "medic-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
